[faet] #30 apple login 구현및 코드 리팩토링
