+++ src/main/java/hanium/dongguk/auth/provider/apple/dto/ApplePublicKey.java
@@ -1,4 +1,4 @@
-package hanium.dongguk.user.patient.auth.dto;
+package hanium.dongguk.auth.provider.apple.dto;
 
 public record ApplePublicKey(
         String kty,
