+++ src/main/java/hanium/dongguk/auth/provider/apple/dto/ApplePublicKeyList.java
@@ -1,0 +1,10 @@
+package hanium.dongguk.auth.provider.apple.dto;
+
+import java.util.List;
+
+public record ApplePublicKeyList(
+
+    List<ApplePublicKey> keyList
+
+){
+}
