+++ src/main/java/hanium/dongguk/global/constants/Constants.java
@@ -9,6 +9,8 @@ public class Constants {
     public static String PREFIX_AUTH = "Authorization";
     public static String ACCESS_COOKIE_NAME = "access_token";
     public static String REFRESH_COOKIE_NAME = "refresh_token";
+    public static String APPLE_BASEURL = "https://appleid.apple.com/auth";
+    public static String GOOGLE_BASEURL = "https://www.googleapis.com/oauth2/v1";
 
     public static List<String> NO_NEED_AUTH = List.of(
             "/swagger",
