+++ src/main/java/hanium/dongguk/user/core/controller/UserController.java
@@ -1,0 +1,10 @@
+package hanium.dongguk.user.core.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UserController implements UserApiSwagger {
+
+}
