--- src/main/java/hanium/dongguk/user/patient/auth/dto/ApplePublicKeysResponseDto.java
@@ -1,10 +1,0 @@
-package hanium.dongguk.user.patient.auth.dto;
-
-import java.util.List;
-
-public record ApplePublicKeysResponseDto (
-
-    List<ApplePublicKey> keyList
-
-){
-}
