--- src/main/java/hanium/dongguk/user/patient/service/KakaoAuthService.java
@@ -1,4 +1,0 @@
-package hanium.dongguk.user.patient.service;
-
-public class KakaoAuthService {
-}
